KAFKA-3896: Fix KStream-KStream leftJoin in RepartitionIntegrationTest
The issue of transiently having duplicates is due to the bad design of the left join itself: in order to ignore the partial joined results such as `A:null`, it lets the producer to potentially send twice to source stream one and rely on all the following conditions to be true in order to pass the test:

1. `receiveMessages` happen to have fetched all the produced results and have committed offsets.
2. streams app happen to have completed sending all result data.
3. consumer used in `receiveMessages` will complete getting all messages in a single poll().

If any of the above is not true, the test fails.

Fixed this test to add a filter right after left join to filter out partial joined results. Minor cleanup on integration test utils.

Author: Guozhang Wang <wangguoz@gmail.com>

Reviewers: Damian Guy, Ewen Cheslack-Postava

Closes #2485 from guozhangwang/K3896-duplicate-join-results