KAFKA-4725; Stop leaking messages in produce request body when requests are delayed
This change is in response to [KAFKA-4725](https://issues.apache.org/jira/browse/KAFKA-4725).

When a produce request is received, if the user/client is exceeding their produce quota, the response will be delayed until the quota is refilled appropriately.

Unfortunately, the request body is still referenced in the callback which in turn leaks the messages contained within the request.

This change allows the `KafkaApis` method to take ownership of the request body from the `RequestChannel.Request` object.

I am not sure whether this breaks other invariants which are assumed within other parts of Kafka.

Author: Tim Carey-Smith <tim@spork.in>

Reviewers: Ismael Juma <ismael@juma.me.uk>, Jun Rao <junrao@gmail.com>

Closes #2496 from halorgium/fix-throttled-response-leak