Use regex for including jars from libs folder instead of adding individual files
When we include individual files to classpath it may exceed the maximum
length of command line which is 4096 bytes.
This issue also breaks the stop script which will grep for
‘kafka.Kafka’ but when cmdline exceeds 4096 bytes ‘kafka.Kafka’ text
won’t be present in the process and hence cannot be grepped.
Details here:
http://unix.stackexchange.com/questions/343353/grep-only-prints-up-to-40
96-characters-of-any-process

Using `CLASSPATH=$CLASSPATH:$base_dir/libs/*` limits the length of
command line.