KAFKA-4789: Added support to ProcessorTopologyTestDriver to forward timestamps to internal topics
This resolves the issue in the ProcessorTopologyTestDriver that the extracted timestamp is not forwarded with the produced record to the internal topics.

JIRA ticket: https://issues.apache.org/jira/browse/KAFKA-4789

The contribution is my original work and I license the work to the project under the project's open source license.

guozhangwang dguy

Author: Hamidreza Afzali <hrafzali@gmail.com>

Reviewers: Damian Guy, Guozhang Wang

Closes #2590 from hrafzali/KAFKA-4789_ProcessorTopologyTestDriver_timestamp