KAFKA-3155; Avoid long overflow in RecordBatch#maybeExpire
`Long.MaxValue` for the linger overflows in `RecordBatch#maybeExpire` when added to
the current timestamp.

Then causes an error to be set for the batch by `Sender` (not happening every time since
it depends on the timing of `Sender`):

That error then causes a call to `ProduceRequestResult#done` on the batch, which then
makes the check for "not done" fail.

Author: Armin Braun <me@obrown.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #2639 from original-brownbear/KAFKA-3155