Use peek to implement print
The functionaliy of KeyValuePrinter is replaced with a printAction that
is being passed to KStreamPeek. We therefore can get rid of
KeyValuePrinter.

KStream.print was extended to handle KeyValueMapper in order to provide
users the option to change output of K and V. Therefore, constructors of
KeyValuePrinter as well as KeyValuePrinterProcessor have to be adapted. The default case
where no mapper will be passed, is still covered by the previous comma
separated out of K, V.