KAFKA-4858: Broker should ignore long topic names
Starting from the 0.10.0.0 release topics names of over 249 characters are not allowed (`kafka-topics.sh` would disallow the creation). However, it is possible to use an older client version against the more recent brokers and attempt to create topics with longer names.
This causes problems with the broker when trying to bring the corresponding partitions online (as described in the [JIRA](https://issues.apache.org/jira/browse/KAFKA-4858)).
This PR makes the broker ignore those topics to avoid the reported failures.