KAFKA-5039: Logging in BlockingChannel and SyncProducer connect
When an exception is thrown in BlockingChannel::connect, the connection is disconnected but the actual exception is not logged. This later manifests as ClosedChannelException when trying to send. Also the SyncProducer wrongfully logs "Connected to host:port for producing" even in case of exceptions.

The proposed patch logs the exception in BlockingChannel::connect in debug level and logs the "Connected" message in SyncProducer only on successful connection.