Fix findbug warnings about OffsetStorageWriter
Based on the description of the class OffsetStorageWriter, it is not a thread-safe class and should be accessed
only from a Task's processing thread. Still many methods within this class have been explicitly synchronized in
their function definiton. The doFlush() method is a non-blocking function and hasn't been synchronized but modifies
the variables used within the synchronized methods in this class. This could lead to potential inconsistent synchronization
of many variables within this class.  We can therefore remove the synchronized keyword from the method signatures within the OffsetStorageWriter class. The WorkerSourceTask class calls the different methods (offset, beginFlush,cancelFlush, handleFinishWrite) within a synchronized block. Hence their method definitions in OffsetStorageWriter.java does not need to contain the 'synchronized' keyword.