KAFKA-5142: Expose Record Headers in Kafka Connect
Add ability to convert the headers data
new ConnectHeader that has header value schema
new SubjectConverter which allows exposing a subject, in this case the subject is the key. - this can be used to register the header type in repos like schema registry 
Add primitive subject converter that allows to configure header types, uses an in memory repo - this could be replace by a remote global repo ala schema repo.