KAFKA-5144: renamed variables in MinTimestampTracker and added comments
The descendingSubsequence is a misnomer. The linked list is actually arranged so that the lowest timestamp is first and larger timestamps are added to the end, therefore renamed to ascendingSubsequence.
The minElem variable was also misnamed. It's actually the current maximum element as it's taken from the end of the list.
Added comment to get() to make it clear it's returning the lowest timestamp.

Author: mihbor <mihbor@users.noreply.github.com>

Reviewers: Matthias J. Sax, Guozhang Wang

Closes #2948 from mihbor/patch-4