MINOR: Fix bug in `waitUntilLeaderIsElectedOrChanged` and simplify result type
Also disable a couple of tests that were passing incorrectly until KAFKA-3096 is fixed.

The bug was for the following case:

`leader.isDefined && oldLeaderOpt.isEmpty && newLeaderOpt.isDefined && newLeaderOpt.get != leader.get`

We would consider it a successful election even though the new leader was not the expected leader.

I also changed the result type as we never return `None` (we throw an exception instead).

Author: Ismael Juma <ismael@juma.me.uk>

Reviewers: Jun Rao <junrao@gmail.com>

Closes #3031 from ijuma/fix-wait-until-leader-is-elected