KAFKA-5291: AdminClient should not trigger auto creation of topics
- Added a boolean allow_auto_topic_creation to MetadataRequest.
I didn’t bump the version a second time since we did it once for this
release already, but this needs to be verified.

- Set it to false in the new AdminClient and StreamsKafkaClient
(which exists for the purpose of creating topics manually); set it
to true everywhere else for now. Other clients will eventually
rely on client-side auto topic creation, but that’s not there yet.