KAFKA-5297: Reduce time required for shutdown
Avoid reopening and recreating page mappings for index files during
truncation unless it's required by the platform.

In my tests this increases throughput of Log#close by about 70%. See
KAFKA-5297 for details.

kafka.log.AbstractIndex#mmap no longer closes the RandomAccessFile, even
if an exception is thrown. I'm not sure if this is the right thing to
do.