KAFKA-5247; Materialize committed offsets in offset order
With this patch, offset commits are always materialized according to the order of the commit records in the offsets topic.

Before this patch, transactional offset commits were materialized in transaction order. However, the log cleaner will always preserve the record with the greatest offset. This meant that if there was a mix of offset commits from a consumer and a transactional producer, then it we would switch from transactional order to offset order after cleaning, resulting in an inconsistent state.

Author: Apurva Mehta <apurva@confluent.io>

Reviewers: Guozhang Wang <wangguoz@gmail.com>, Jason Gustafson <jason@confluent.io>

Closes #3108 from apurvam/KAFKA-5247-materialize-committed-offsets-in-offset-order

(cherry picked from commit d1853f791147d66303c2700a8e58af762518188c)
Signed-off-by: Jason Gustafson <jason@confluent.io>