KAFKA-5273: Make KafkaConsumer.committed query the server for all partitions
Before this patch the consumer would return the cached offsets for partitions in its current assignment. This worked when all the offset commits went through the consumer.

With KIP-98, offsets can be committed transactionally through the producer. This means that relying on cached positions in the consumer returns incorrect information: since commits go through the producer, the cache is never updated.

Hence we need to update the `KafkaConsumer.committed` method to always lookup the server for the last committed offset to ensure it gets the correct information every time.

Author: Apurva Mehta <apurva@confluent.io>

Reviewers: Jason Gustafson, Guozhang Wang

Closes #3119 from apurvam/KAFKA-5273-kafkaconsumer-committed-should-always-hit-server