MINOR: fix flakiness in testDeleteAcls
This call to isCompletedExceptionally introduced a race condition
because the future might not have been completed.  assertFutureError
checks that the exception is present and of the correct type in any
case, so the call was not necessary.

Author: Colin P. Mccabe <cmccabe@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #3139 from cmccabe/fix-test-deleteacls

(cherry picked from commit a10990f44b1a4f8345eae7ebfc9c571bfd0447c3)
Signed-off-by: Ismael Juma <ismael@juma.me.uk>