KAFKA-5147; Add missing synchronization to TransactionManager
The basic idea is that exactly three collections, ie. `pendingRequests`, `newPartitionsToBeAddedToTransaction`, and `partitionsInTransaction` are accessed from the context of application threads. The first two are modified from the application threads, and the last is read from those threads.

So to make the `TransactionManager` truly thread safe, we have to ensure that all accesses to these three members are done in a synchronized block. I inspected the code, and I believe this patch puts the synchronization in all the correct places.

Author: Apurva Mehta <apurva@confluent.io>

Reviewers: Jason Gustafson <jason@confluent.io>

Closes #3132 from apurvam/KAFKA-5147-transaction-manager-synchronization-fixes

(cherry picked from commit 02c0c3b01730bdbff8a09d355c1b017715c7ce10)
Signed-off-by: Jason Gustafson <jason@confluent.io>