KAFKA-5164: Ensure SetSchemaMetadata updates key or value when Schema changes
When the `SetSchemaMetadata` SMT is used to change the name and/or version of the key or value’s schema, any references to the old schema in the key or value must be changed to reference the new schema. Only keys or values that are `Struct` have such references, and so currently only these are adjusted.

This is based on `trunk` since the fix is expected to be targeted to the 0.11.1 release.

Author: Randall Hauch <rhauch@gmail.com>

Reviewers: Ewen Cheslack-Postava <ewen@confluent.io>

Closes #3198 from rhauch/kafka-5164

(cherry picked from commit af85e05b98a41cd5f8ac45a853b2ddd28463c084)
Signed-off-by: Ewen Cheslack-Postava <me@ewencp.org>